# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Tower Go 用户管理系统 API 文档",
        "title": "Tower Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "jeckwell@126.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:10024",
    "basePath": "/api/v1",
    "paths": {
...
        "/store-accounts": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "门店记账"
                ],
                "summary": "记账列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "门店ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "商品ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "销售渠道",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "订单来源",
                        "name": "order_source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "订单编号",
                        "name": "order_no",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.StoreAccount"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "门店记账"
                ],
                "summary": "创建记账",
                "parameters": [
                    {
                        "description": "记账信息",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateStoreAccountReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.StoreAccount"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/store-accounts/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "门店记账"
                ],
                "summary": "记账统计",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "门店ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/store-accounts/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "门店记账"
                ],
                "summary": "获取记账详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "记账ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.StoreAccount"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "门店记账"
                ],
                "summary": "更新记账",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "记账ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "记账信息",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateStoreAccountReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "门店记账"
                ],
                "summary": "删除记账",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "记账ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateStoreAccountReq": {
            "type": "object",
            "required": [
                "channel",
                "order_source",
                "product_id",
                "quantity"
            ],
            "properties": {
                "account_date": {
                    "description": "格式: 2024-12-07",
                    "type": "string"
                },
                "amount": {
                    "type": "number",
                    "minimum": 0
                },
                "channel": {
                    "type": "string",
                    "maxLength": 50
                },
                "order_no": {
                    "type": "string",
                    "maxLength": 100
                },
                "order_source": {
                    "type": "string",
                    "maxLength": 50
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "remark": {
                    "type": "string",
                    "maxLength": 500
                },
                "spec": {
                    "type": "string",
                    "maxLength": 100
                },
                "unit": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "model.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "角色代码：admin(总部管理员), store_admin(门店管理员), staff(普通员工)",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "角色描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "status": {
                    "description": "状态：1=启用 0=禁用",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Store": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "门店地址",
                    "type": "string"
                },
                "business_hours": {
                    "description": "营业时间",
                    "type": "string"
                },
                "contact_person": {
                    "description": "联系人",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "门店名称",
                    "type": "string"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "状态：1=正常，2=停业",
                    "type": "integer"
                },
                "store_code": {
                    "description": "门店编码 JWXXXX",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.StoreAccount": {
            "type": "object",
            "properties": {
                "account_date": {
                    "type": "string"
                },
                "account_no": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "channel": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "operator": {
                    "$ref": "#/definitions/model.User"
                },
                "operator_id": {
                    "type": "integer"
                },
                "order_no": {
                    "type": "string"
                },
                "order_source": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product": {
                    "$ref": "#/definitions/model.SupplierProduct"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "remark": {
                    "type": "string"
                },
                "spec": {
                    "type": "string"
                },
                "store": {
                    "description": "关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Store"
                        }
                    ]
                },
                "store_id": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Supplier": {
            "type": "object",
            "properties": {
                "contact_email": {
                    "type": "string"
                },
                "contact_person": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "supplier_address": {
                    "type": "string"
                },
                "supplier_code": {
                    "type": "string"
                },
                "supplier_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SupplierCategory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "supplier": {
                    "$ref": "#/definitions/model.Supplier"
                },
                "supplier_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SupplierProduct": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/model.SupplierCategory"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "remark": {
                    "type": "string"
                },
                "spec": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "supplier": {
                    "$ref": "#/definitions/model.Supplier"
                },
                "supplier_id": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateStoreAccountReq": {
            "type": "object",
            "properties": {
                "account_date": {
                    "type": "string"
                },
                "amount": {
                    "type": "number",
                    "minimum": 0
                },
                "channel": {
                    "type": "string",
                    "maxLength": 50
                },
                "order_no": {
                    "type": "string",
                    "maxLength": 100
                },
                "order_source": {
                    "type": "string",
                    "maxLength": 50
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "remark": {
                    "type": "string",
                    "maxLength": 500
                },
                "spec": {
                    "type": "string",
                    "maxLength": 100
                },
                "unit": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "employee_no": {
                    "description": "工号，6位数字",
                    "type": "string"
                },
                "gender": {
                    "description": "性别：1=男，2=女",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_login_at": {
                    "description": "记录最后登录时间（仅后端维护）",
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "角色关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Role"
                        }
                    ]
                },
                "role_id": {
                    "description": "--- 权限管理 (RBAC) ---",
                    "type": "integer"
                },
                "status": {
                    "description": "--- 状态与安全管理 ---",
                    "type": "integer"
                },
                "store": {
                    "description": "门店关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Store"
                        }
                    ]
                },
                "store_id": {
                    "description": "--- 门店关联 ---",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}